issue #10: fix crash when acking messages on normal queues
The `ack` callback was expecting the ack tags to be always
encapsulated within a `dqack` record which was not the case.

Several callbacks, in fact, delegate their logic to the passthrough
functions which do not handle the returned values but just the queue
state.

This patch simplifies the `ack` logic dealing with `dqacks` only when
the queue is a deduplication queue and acknowledgement is enabled by
the channels. For all the other cases, it falls back to the passthrough
logic.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>