issues #16, #20: fix cache supervisor rabbit_boot_step
The cache supervisor boot step was starting the process directly via
the `start_link` function. This is error prone as we are not sure
about the life-cycle of the bootstrap process.

It appeared to work during a proper bootstrap of the broker but was
faulty when enabling the plugin on a running one.

This patch starts the cache supervisor as a child of the main rabbit
supervisor process correctly ensuring its management.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>