issue #22: fix `publish_delivered` callback with deduplication queue
When the queue is empty, the message is passed straight to the first
consumer to save one round trip time.

In such cases, if message acknowledgement is enabled,
the `publish_delivered` callback is called instead of `fetch`.

If de-duplication is enabled, we need to ensure a `dqack`
is returned instead of a simple acknowledgement tag.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>