issue #36: fix queue crash when deleting it
When deleting the queue via the Management console, few queue
information requests are still sent to the deleted queue.

As the queue does not exist anymore, so does the cache. Therefore,
we ignore errors during deleted cache info lookups.

This could not be catched via tests as the behaviour is triggered by
the management plugin.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>