issue #49, #55, #58: ensure mnesia cache persistence over node failures
The deduplication cache was created on the node where the
resource (exchange or queue) was initialized.

If the node crashed and was replaced mnesia could not recover or
recreate the table used as a cache.

After several attempts, it seems the best way to prevent this from
happening is to replicate the table over more nodes to ensure mnesia
can recover it in case of a node disappearing.

To ensure safe enough persistence without replicating too
aggressively, the cache is created on 2/3 of the total cluster nodes.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>