queue: remove cache persistence
RabbitMQ classic queue supports both transient and persistent messages
on the same queue. If a node where a queue relies crashes, transient
messages will be lost whereas persistent will remain.

It is unfortunately non-trivial from the plugin to access to
presistent messages at recover. On the other side, persisting
deduplication headers would be dangerous as we might deduplicate lost
transient messages forever in case of recovery.

The easiest solution is, for now, to flush the cache at startup.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>