issue #62: drop RMQ boot steps in favour of application behaviour
RabbitMQ plugins are instantiated at the end of the broker boot
routing. Therefore, their subscription to boot steps is considered way
after the same have been executed.

This causes problems when restarting the broker with deduplication
queues as all queues are initialized before the
`rabbit_message_deduplication_queue` module has been set as
`backing_queue_module`. The end result is all deduplication queues
are believed to be normal ones.

As boot steps are not functional for plugins initialization, we switch
to standard Erlang/Elixir application behaviour.

More information in this discussion:
https://groups.google.com/g/rabbitmq-users/c/m7jgGmIk4oY

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>