Improve cache creation logic resilience
When loading a Mnesia table, there might be chances the application
waits forever. This might be due, for example, to the cluster going
offline and resuming each node in a specificly wrong order.

For example if we have two nodes and we restart them in rapid sequence,
the first node might timeout reloading its tables.
This is due to the fact the first node did not know the other
one is rebooting and it waits for updates on the tables.

This is particularly important if we create a table on a single node
and we then replace it (node and persistent disk). The new node taking
its place will wait forever for a table which belonged only to its
predecessor.

The solution consist of forcing the loading of a Mnesia table if we
timeout waiting for it. We sacrifice data consistency for the sake of
better availability.

Also, fixing a bug in which the table creation timeout was not in sync
across modules.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>