Enable optional cache replication
As RMQ queues are local and single-process, it makes little sense to
replicate their deduplication caches on multiple nodes.

High availability is, in this case, undesirable as it increases the
pressure on the cache without providing benefits. If the node on which
the queue crashes, the queue becomes unavailable.

Moreover, in case of transient messages and queue crash, we would
persist the deduplication headers undefinitely.

The new logic creates local caches on the nodes where the queues
reside. Thus, reducing the overhead of copying the cache entries on multiple
nodes.

If the node crashes, we loose the content of the deduplication
cache. In future, we might recover the headers from the persistent
messages.

For exchanges deduplication we keep replicating the caches on 2/3 of
the cluster nodes.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>