cache: query list of nodes from Mnesia
Previously, we were retrieving the list of nodes onto which to
replicate the cache by querying the VM. This had the side effect of
retrieving all connected nodes and not only those in which Mnesia
schema was installed.

This is usually not an issue as it's unlikely someone would connect a
BEAM VM to RabbitMQ and the broker installs the schema on all its
running nodes.

Nevertheless, the system tests are an exception to the above
assumption and they where failing when trying to replicate a disc
cache onto the test VM.

To remediate that, we now retrieve the list of nodes by querying
Mnesia.system_info rather than Erlang.Node.

An alternative would be to query the list of nodes through RabbitMQ
itself but I'd rather keep the caching logic RMQ agnostic.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>