issue #82: fully implement cache replication logic
When a new node joins the Mnesia cluster, the list of connected nodes
is compared with the list of nodes where each cache is defined within
the DB schema.

If the 2/3 ratio is not met anymore, the cache is replicated onto the
subsequent nodes.

This logic does not cover the case where ad node is down. Therefore,
if a cluster was to rotate its nodes, the previous nodes should be
removed from the DB schema. Otherwise, the cache will still be
considered as replicated on the old node.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>