issue #97: migrate old caches to new configuration
Cache options have been changed in 0.6.0 without considering old
exising ones on running brokers.

This should not be a big issue for queues which are re-initialized at
each broker restart. Nevertheless, the same does not happen for
exchanges which are set up once and their configuration persists over
reboots.

The logic now inspects all existing caches and reconfigures those
which appear to have old configurations.

The logic is imperfect as it cannot properly handle situations where
exchanges where created on a single node cluster and the cluster is
then scaled up. Nevertheless, this should be pretty much a corner case.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>