issue #97: rework eschange creation logic
Latest version of RabbitMQ changed the `exchange_type` behavior (so
much for backwards compatibility).

While inspecting the code, it turns out that the exhange_type:create
is called upon recovery as well. This simplifies the logic as we can
assume that old caches will be migrated on restart (needs to be tested).

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>