fix: do not rely on `config/confix.exs` when building with `erlang.mk`
It seems recent version of `erlang.mk` now supports Elixir as
language. Unfortunately (or brilliantly?) it insists in generating its
own `.app` file instead of using `mix.exs`.

As a result, incomplete `.app` file is generated missing important
details such as the environment configuration stored within
`config/config.exs`.

As RMQ build logic is the usual nightmare, instead of patching the
nonsense, we opt to resolve the issue within the module itself. If the
Application environment is not populated, we default the values within
the app code.

This allows RMQ convoluted build logic altogether with `erlang.mk` to
generate the incomplete `.app` file which will be used when using
commands such as `make run-broker` or `make tests`.