fix(queue): insertion and duplication test into separate callbacks
The logic used to insert and test for cache membership at the same
time. This to protect against race conditions.

Since classic queues run in single processes within a single node,
there is no such risk for race condition.

The side effect of previous implementation was the challenge in
dealing with callbacks which tested for membership but did not publish
the message in the queue (issue #96).

The new implementation treats cache membership testing and insertion
in the right callbacks: `is_duplicate` and `publish`.

issue #118.